+++ test/transform/resource/after-ecj/LoggerFloggerAccess.java
@@ -1,0 +1,52 @@
+import lombok.AccessLevel;
+import lombok.extern.flogger.Flogger;
+@Flogger(access = AccessLevel.PUBLIC) class LoggerFloggerAccessPublic {
+  public static final @java.lang.SuppressWarnings("all") @lombok.Generated com.google.common.flogger.FluentLogger log = com.google.common.flogger.FluentLogger.forEnclosingClass();
+  <clinit>() {
+  }
+
+  LoggerFloggerAccessPublic() {
+    super();
+  }
+}
+@Flogger(access = AccessLevel.MODULE) class LoggerFloggerAccessModule {
+  static final @java.lang.SuppressWarnings("all") @lombok.Generated com.google.common.flogger.FluentLogger log = com.google.common.flogger.FluentLogger.forEnclosingClass();
+  <clinit>() {
+  }
+
+  LoggerFloggerAccessModule() {
+    super();
+  }
+}
+@Flogger(access = AccessLevel.PROTECTED) class LoggerFloggerAccessProtected {
+  protected static final @java.lang.SuppressWarnings("all") @lombok.Generated com.google.common.flogger.FluentLogger log = com.google.common.flogger.FluentLogger.forEnclosingClass();
+  <clinit>() {
+  }
+
+  LoggerFloggerAccessProtected() {
+    super();
+  }
+}
+@Flogger(access = AccessLevel.PACKAGE) class LoggerFloggerAccessPackage {
+  static final @java.lang.SuppressWarnings("all") @lombok.Generated com.google.common.flogger.FluentLogger log = com.google.common.flogger.FluentLogger.forEnclosingClass();
+  <clinit>() {
+  }
+
+  LoggerFloggerAccessPackage() {
+    super();
+  }
+}
+@Flogger(access = AccessLevel.PRIVATE) class LoggerFloggerAccessPrivate {
+  private static final @java.lang.SuppressWarnings("all") @lombok.Generated com.google.common.flogger.FluentLogger log = com.google.common.flogger.FluentLogger.forEnclosingClass();
+  <clinit>() {
+  }
+
+  LoggerFloggerAccessPrivate() {
+    super();
+  }
+}
+@Flogger(access = AccessLevel.NONE) class LoggerFloggerAccessNone {
+  LoggerFloggerAccessNone() {
+    super();
+  }
+}