+++ test/transform/resource/after-ecj/LoggerJBossLogAccess.java
@@ -1,0 +1,52 @@
+import lombok.AccessLevel;
+import lombok.extern.jbosslog.JBossLog;
+@JBossLog(access = AccessLevel.PUBLIC) class LoggerJBossLogAccessPublic {
+  public static final @java.lang.SuppressWarnings("all") @lombok.Generated org.jboss.logging.Logger log = org.jboss.logging.Logger.getLogger(LoggerJBossLogAccessPublic.class);
+  <clinit>() {
+  }
+
+  LoggerJBossLogAccessPublic() {
+    super();
+  }
+}
+@JBossLog(access = AccessLevel.MODULE) class LoggerJBossLogAccessModule {
+  static final @java.lang.SuppressWarnings("all") @lombok.Generated org.jboss.logging.Logger log = org.jboss.logging.Logger.getLogger(LoggerJBossLogAccessModule.class);
+  <clinit>() {
+  }
+
+  LoggerJBossLogAccessModule() {
+    super();
+  }
+}
+@JBossLog(access = AccessLevel.PROTECTED) class LoggerJBossLogAccessProtected {
+  protected static final @java.lang.SuppressWarnings("all") @lombok.Generated org.jboss.logging.Logger log = org.jboss.logging.Logger.getLogger(LoggerJBossLogAccessProtected.class);
+  <clinit>() {
+  }
+
+  LoggerJBossLogAccessProtected() {
+    super();
+  }
+}
+@JBossLog(access = AccessLevel.PACKAGE) class LoggerJBossLogAccessPackage {
+  static final @java.lang.SuppressWarnings("all") @lombok.Generated org.jboss.logging.Logger log = org.jboss.logging.Logger.getLogger(LoggerJBossLogAccessPackage.class);
+  <clinit>() {
+  }
+
+  LoggerJBossLogAccessPackage() {
+    super();
+  }
+}
+@JBossLog(access = AccessLevel.PRIVATE) class LoggerJBossLogAccessPrivate {
+  private static final @java.lang.SuppressWarnings("all") @lombok.Generated org.jboss.logging.Logger log = org.jboss.logging.Logger.getLogger(LoggerJBossLogAccessPrivate.class);
+  <clinit>() {
+  }
+
+  LoggerJBossLogAccessPrivate() {
+    super();
+  }
+}
+@JBossLog(access = AccessLevel.NONE) class LoggerJBossLogAccessNone {
+  LoggerJBossLogAccessNone() {
+    super();
+  }
+}