+++ test/transform/resource/after-ecj/LoggerLog4jAccess.java
@@ -1,0 +1,52 @@
+import lombok.AccessLevel;
+import lombok.extern.log4j.Log4j;
+@Log4j(access = AccessLevel.PUBLIC) class LoggerLog4jAccessPublic {
+  public static final @java.lang.SuppressWarnings("all") @lombok.Generated org.apache.log4j.Logger log = org.apache.log4j.Logger.getLogger(LoggerLog4jAccessPublic.class);
+  <clinit>() {
+  }
+
+  LoggerLog4jAccessPublic() {
+    super();
+  }
+}
+@Log4j(access = AccessLevel.MODULE) class LoggerLog4jAccessModule {
+  static final @java.lang.SuppressWarnings("all") @lombok.Generated org.apache.log4j.Logger log = org.apache.log4j.Logger.getLogger(LoggerLog4jAccessModule.class);
+  <clinit>() {
+  }
+
+  LoggerLog4jAccessModule() {
+    super();
+  }
+}
+@Log4j(access = AccessLevel.PROTECTED) class LoggerLog4jAccessProtected {
+  protected static final @java.lang.SuppressWarnings("all") @lombok.Generated org.apache.log4j.Logger log = org.apache.log4j.Logger.getLogger(LoggerLog4jAccessProtected.class);
+  <clinit>() {
+  }
+
+  LoggerLog4jAccessProtected() {
+    super();
+  }
+}
+@Log4j(access = AccessLevel.PACKAGE) class LoggerLog4jAccessPackage {
+  static final @java.lang.SuppressWarnings("all") @lombok.Generated org.apache.log4j.Logger log = org.apache.log4j.Logger.getLogger(LoggerLog4jAccessPackage.class);
+  <clinit>() {
+  }
+
+  LoggerLog4jAccessPackage() {
+    super();
+  }
+}
+@Log4j(access = AccessLevel.PRIVATE) class LoggerLog4jAccessPrivate {
+  private static final @java.lang.SuppressWarnings("all") @lombok.Generated org.apache.log4j.Logger log = org.apache.log4j.Logger.getLogger(LoggerLog4jAccessPrivate.class);
+  <clinit>() {
+  }
+
+  LoggerLog4jAccessPrivate() {
+    super();
+  }
+}
+@Log4j(access = AccessLevel.NONE) class LoggerLog4jAccessNone {
+  LoggerLog4jAccessNone() {
+    super();
+  }
+}