+++ test/transform/resource/after-ecj/LoggerSlf4jAccess.java
@@ -1,0 +1,52 @@
+import lombok.AccessLevel;
+import lombok.extern.slf4j.Slf4j;
+@Slf4j(access = AccessLevel.PUBLIC) class LoggerSlf4jAccessPublic {
+  public static final @java.lang.SuppressWarnings("all") @lombok.Generated org.slf4j.Logger log = org.slf4j.LoggerFactory.getLogger(LoggerSlf4jAccessPublic.class);
+  <clinit>() {
+  }
+
+  LoggerSlf4jAccessPublic() {
+    super();
+  }
+}
+@Slf4j(access = AccessLevel.MODULE) class LoggerSlf4jAccessModule {
+  static final @java.lang.SuppressWarnings("all") @lombok.Generated org.slf4j.Logger log = org.slf4j.LoggerFactory.getLogger(LoggerSlf4jAccessModule.class);
+  <clinit>() {
+  }
+
+  LoggerSlf4jAccessModule() {
+    super();
+  }
+}
+@Slf4j(access = AccessLevel.PROTECTED) class LoggerSlf4jAccessProtected {
+  protected static final @java.lang.SuppressWarnings("all") @lombok.Generated org.slf4j.Logger log = org.slf4j.LoggerFactory.getLogger(LoggerSlf4jAccessProtected.class);
+  <clinit>() {
+  }
+
+  LoggerSlf4jAccessProtected() {
+    super();
+  }
+}
+@Slf4j(access = AccessLevel.PACKAGE) class LoggerSlf4jAccessPackage {
+  static final @java.lang.SuppressWarnings("all") @lombok.Generated org.slf4j.Logger log = org.slf4j.LoggerFactory.getLogger(LoggerSlf4jAccessPackage.class);
+  <clinit>() {
+  }
+
+  LoggerSlf4jAccessPackage() {
+    super();
+  }
+}
+@Slf4j(access = AccessLevel.PRIVATE) class LoggerSlf4jAccessPrivate {
+  private static final @java.lang.SuppressWarnings("all") @lombok.Generated org.slf4j.Logger log = org.slf4j.LoggerFactory.getLogger(LoggerSlf4jAccessPrivate.class);
+  <clinit>() {
+  }
+
+  LoggerSlf4jAccessPrivate() {
+    super();
+  }
+}
+@Slf4j(access = AccessLevel.NONE) class LoggerSlf4jAccessNone {
+  LoggerSlf4jAccessNone() {
+    super();
+  }
+}