+++ test/transform/resource/after-ecj/LoggerXslf4jAccess.java
@@ -1,0 +1,52 @@
+import lombok.AccessLevel;
+import lombok.extern.slf4j.XSlf4j;
+@XSlf4j(access = AccessLevel.PUBLIC) class LoggerXslf4jAccessPublic {
+  public static final @java.lang.SuppressWarnings("all") @lombok.Generated org.slf4j.ext.XLogger log = org.slf4j.ext.XLoggerFactory.getXLogger(LoggerXslf4jAccessPublic.class);
+  <clinit>() {
+  }
+
+  LoggerXslf4jAccessPublic() {
+    super();
+  }
+}
+@XSlf4j(access = AccessLevel.MODULE) class LoggerXslf4jAccessModule {
+  static final @java.lang.SuppressWarnings("all") @lombok.Generated org.slf4j.ext.XLogger log = org.slf4j.ext.XLoggerFactory.getXLogger(LoggerXslf4jAccessModule.class);
+  <clinit>() {
+  }
+
+  LoggerXslf4jAccessModule() {
+    super();
+  }
+}
+@XSlf4j(access = AccessLevel.PROTECTED) class LoggerXslf4jAccessProtected {
+  protected static final @java.lang.SuppressWarnings("all") @lombok.Generated org.slf4j.ext.XLogger log = org.slf4j.ext.XLoggerFactory.getXLogger(LoggerXslf4jAccessProtected.class);
+  <clinit>() {
+  }
+
+  LoggerXslf4jAccessProtected() {
+    super();
+  }
+}
+@XSlf4j(access = AccessLevel.PACKAGE) class LoggerXslf4jAccessPackage {
+  static final @java.lang.SuppressWarnings("all") @lombok.Generated org.slf4j.ext.XLogger log = org.slf4j.ext.XLoggerFactory.getXLogger(LoggerXslf4jAccessPackage.class);
+  <clinit>() {
+  }
+
+  LoggerXslf4jAccessPackage() {
+    super();
+  }
+}
+@XSlf4j(access = AccessLevel.PRIVATE) class LoggerXslf4jAccessPrivate {
+  private static final @java.lang.SuppressWarnings("all") @lombok.Generated org.slf4j.ext.XLogger log = org.slf4j.ext.XLoggerFactory.getXLogger(LoggerXslf4jAccessPrivate.class);
+  <clinit>() {
+  }
+
+  LoggerXslf4jAccessPrivate() {
+    super();
+  }
+}
+@XSlf4j(access = AccessLevel.NONE) class LoggerXslf4jAccessNone {
+  LoggerXslf4jAccessNone() {
+    super();
+  }
+}